feat: Add FieldToolbar to expression fields
