+++ Composer/packages/lib/code-editor/src/components/FieldToolbar/__tests__/ToolbarButtonMenu.test.tsx
@@ -8,7 +8,7 @@ import { act, fireEvent, render, screen } from '@botframework-composer/test-util
 import React from 'react';
 
 import { ToolbarButtonMenu } from '../ToolbarButtonMenu';
-import { FunctionRefPayload, PropertyRefPayload, TemplateRefPayload } from '../types';
+import { FunctionRefPayload, PropertyRefPayload, TemplateRefPayload } from '../../../lg/types';
 
 (global as any).crypto = {
   getRandomValues: (arr: any[]) => crypto.randomBytes(arr.length),
