+++ Composer/packages/lib/code-editor/src/components/FieldToolbar/index.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export * from './FieldToolbar';
