+++ Composer/packages/lib/code-editor/src/types.ts
@@ -42,3 +42,32 @@ export type LgCodeEditorProps = LgCommonEditorProps &
         }
       | string;
   };
+
+export type PropertyItem = {
+  id: string;
+  name: string;
+  children: PropertyItem[];
+};
+
+export type TemplateRefPayload = {
+  kind: 'template';
+  data: {
+    templates: readonly LgTemplate[];
+    onSelectTemplate: (templateString: string, itemType: 'template') => void;
+  };
+};
+
+export type PropertyRefPayload = {
+  kind: 'property';
+  data: { properties: readonly string[]; onSelectProperty: (property: string, itemType: 'property') => void };
+};
+
+export type FunctionRefPayload = {
+  kind: 'function';
+  data: {
+    functions: readonly { key: string; name: string; children: string[] }[];
+    onSelectFunction: (functionString: string, itemType: 'function') => void;
+  };
+};
+
+export type ToolbarButtonPayload = TemplateRefPayload | PropertyRefPayload | FunctionRefPayload;
